flask/requirements.txt
Flask==1.1.2
flask-cors==3.0.10
pytest==6.0.1
pytest-cov==2.10.1
gunicorn==20.0.4
html2text==2020.1.16
sentence_transformers==0.4.1.2
pandas==0.25.3
torch==1.6.0
scikit-learn==0.21.2
spacy==2.3.5
faiss-cpu==1.6.5
Flask-RESTful==0.3.8
Unidecode==1.1.2
newspaper3k==0.2.8
deeppavlov==0.14.0
firebase-admin==2.13.0

